+++ executor/src/witgen/data_structures/mod.rs
@@ -1,2 +1,3 @@
 pub mod column_map;
+pub mod copy_constraints;
 pub mod finalizable_data;
